eth_monitor/requirements.txt
websocket-client==1.6.4
requests==2.31.0
sqlalchemy==2.0.23
python-decouple==3.8
aiohttp==3.9.1
asyncio-mqtt==0.16.1
pydantic==2.5.0
cryptography==41.0.8
confluent-kafka==2.3.0 